requirements.txt
# Core ML/DL frameworks
torch>=2.0.0
pytorch-lightning>=2.0.0
torchvision>=0.15.0

# Computer vision and image processing
opencv-python>=4.8.0
mediapipe>=0.10.0

# Signal processing
PyWavelets>=1.4.0

# Machine learning utilities
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Configuration management
PyYAML>=6.0
omegaconf>=2.3.0

# Experiment tracking and visualization
wandb>=0.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Performance and optimization
tqdm>=4.65.0
psutil>=5.9.0

# Data handling
h5py>=3.9.0
pillow>=10.0.0